requirements.txt
# Core AI and LangChain dependencies
langchain>=0.1.0
langgraph>=0.0.26
langchain-google-genai>=0.0.8
pydantic>=2.5.0

# Telegram Bot
python-telegram-bot>=20.0

# HTTP requests
requests>=2.31.0

# Google APIs
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.111.0

# Environment variables
python-dotenv>=1.0.0

# Date/time handling
python-dateutil>=2.8.0

# Async support
aiohttp>=3.9.0 